add test keys, move to folder of feature and move common function and constants to helper file
